Use fib_lookup on chain selector
What is missing in this patch is to remove the fib_lookup at the end of the ingress pipeline,
since it is done on the ingress part.
Moreover, it could be adopted on the egress part but after a very small test I was not able
to make it work.